+++ src/markerparts/affliationdimension.js
@@ -35,5 +35,5 @@ module.exports = function affliationdimension(){
 	}
 	//outline
 	if (this.outlineWidth > 0) drawArray1.push(MS.outline(drawArray2, this.outlineWidth, this.strokeWidth, this.outlineColor));
-	return MS.buildingBlock(drawArray1,drawArray2,bbox);
+	return {pre: drawArray1, post: drawArray2, bbox: bbox};
 }