+++ src/markerparts/textfields.js
@@ -88,7 +88,7 @@ module.exports = function textfields(){
 			
 			//outline
 			if (this.outlineWidth > 0) drawArray1.push(MS.outline(drawArray2, this.outlineWidth, this.strokeWidth, this.outlineColor))
-			return MS.buildingBlock(drawArray1,drawArray2,gbbox);
+			return {pre: drawArray1, post: drawArray2, bbox: gbbox};
 		}			
 	}
 	
@@ -268,5 +268,5 @@ module.exports = function textfields(){
 		if (this.outlineWidth > 0) drawArray1.push(MS.outline(drawArray2, this.outlineWidth, this.strokeWidth, this.outlineColor))
 
 	}
-	return MS.buildingBlock(drawArray1,drawArray2,gbbox );
+	return {pre: drawArray1, post: drawArray2, bbox: gbbox};
 }